Change array's data type

+++ stanDevi.cpp
@@ -6,7 +6,7 @@
 using namespace std;
 
 
-double stanDevi(int array[], int arraySize) {
+double stanDevi(double array[], int arraySize) {
 
 
     double variStan = variance(array, arraySize);
